requirements.txt
# FastAPI framework and related dependencies
fastapi==0.115.4
uvicorn==0.23.1
starlette==0.41.2
aiofiles==23.1.0
jinja2==3.1.3

# LLM and agent-related libraries
ollama==0.3.3
autogen-agentchat==0.2.37
autogen==0.3.1
sentence-transformers==3.2.1
chromadb==0.5.15

# Utilities and tooling
portalocker==2.10.1
pypdf==5.0.1
flaml[automl]==2.3.1
ruff==0.7.0

ipython==8.29.0
markdownify==0.13.1
requests==2.32.3
pdfplumber==0.11.4
urllib3==2.2.3
colorama==0.4.6

# SQLAlchemy for database ORM
SQLAlchemy==2.0.36
alembic==1.12.0

python-multipart==0.0.16
neo4j==5.8.0
python-dotenv==1.0.1

itsdangerous==2.2.0
rich==13.9.4

nltk